+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state.dart
@@ -307,6 +307,10 @@ abstract class MPTH2FileEditState {
         th2FileEditController.splitMergeController
             .prepareSplitLineAtSelectedEndPoints();
         return true;
+      case MPButtonType.splitLinesAtCrossings:
+        th2FileEditController.splitMergeController
+            .prepareSplitLinesAtCrossings();
+        return true;
       case MPButtonType.undo:
         th2FileEditController.undo();
         return true;