+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_select_non_empty_selection.dart
@@ -362,7 +362,14 @@ class MPTH2FileEditStateSelectNonEmptySelection extends MPTH2FileEditState
           keyProcessed = true;
         }
       case LogicalKeyboardKey.keyX:
-        if ((isCtrlPressed || isMetaPressed) && !isShiftPressed) {
+        if ((isCtrlPressed || isMetaPressed) &&
+            isShiftPressed &&
+            !isAltPressed) {
+          th2FileEditController.stateController.onButtonPressed(
+            MPButtonType.splitLinesAtCrossings,
+          );
+          keyProcessed = true;
+        } else if ((isCtrlPressed || isMetaPressed) && !isShiftPressed) {
           if (isAltPressed) {
             th2FileEditController.areaLineCreationController
                 .createMapConnectionLines();